Moved queue logic into separate file and class
